+++ Pub/PublishApps.ps1
@@ -33,6 +33,10 @@ Remove-Item "$packagesRootDir/ReleaseNote.txt" -ErrorAction Ignore;
 & "$solutionDir/VpnHood.Client.Device.WinDivert/_publish.ps1";
 & "$solutionDir/VpnHood.Client.App/_publish.ps1";
 & "$solutionDir/VpnHood.Client.App.WebServer/_publish.ps1";
+& "$solutionDir/VpnHood.Client.App.Store/_publish.ps1";
+& "$solutionDir/VpnHood.Client.App.Android.Common/_publish.ps1";
+& "$solutionDir/VpnHood.Client.App.Android.GooglePlay/_publish.ps1";
+& "$solutionDir/VpnHood.Client.App.Android.GooglePlay.Core/_publish.ps1";
 
 & "$solutionDir/VpnHood.Server/_publish.ps1";
 & "$solutionDir/VpnHood.Server.Access/_publish.ps1";
@@ -55,13 +59,14 @@ if ($server)
 # publish android
 if ($android)
 {	
-	& "$solutionDir/VpnHood.Client.App.Android/_publish.ps1" -apk;
+	& "$solutionDir/VpnHood.Client.App.Android/_publish_aab.ps1";
+	& "$solutionDir/VpnHood.Client.App.Android/_publish_apk.ps1";
 }
 
 # publish android
 if ($androidConnect)
 {	
-	& "$solutionDir/VpnHood.Client.App.Android.connect/_publish.ps1" -apk;
+	& "$solutionDir/VpnHood.Client.App.Android.connect/_publish_aab.ps1";
 }