+++ VpnHood.Client.App.Android.Connect/MainActivity.cs
@@ -1,9 +1,13 @@
+using System.Diagnostics.CodeAnalysis;
 using Android.Content;
 using Android.Content.PM;
-using Android.Gms.Ads;
 using Android.Service.QuickSettings;
 using Android.Views;
+using VpnHood.Client.App.Abstractions;
 using VpnHood.Client.App.Droid.Common.Activities;
+using VpnHood.Client.App.Droid.Connect.Properties;
+using VpnHood.Client.App.Droid.GooglePlay;
+using VpnHood.Client.App.Store;
 
 namespace VpnHood.Client.App.Droid.Connect;
 
@@ -21,13 +25,32 @@ namespace VpnHood.Client.App.Droid.Connect;
 
 [IntentFilter([Intent.ActionMain], Categories = [Intent.CategoryLauncher, Intent.CategoryLeanbackLauncher])]
 [IntentFilter([TileService.ActionQsTilePreferences])]
-
+[SuppressMessage("ReSharper", "UnusedMember.Global")]
 public class MainActivity : AndroidAppWebViewMainActivity
 {
+    protected override bool ListenToAllIps => AssemblyInfo.ListenToAllIps;
+    protected override int? DefaultSpaPort => AssemblyInfo.DefaultSpaPort;
+
+    protected override IAppUpdaterService CreateAppUpdaterService()
+    {
+        return new GooglePlayAppUpdaterService(this);
+    }
+
     protected override void OnCreate(Bundle? savedInstanceState)
     {
         base.OnCreate(savedInstanceState);
-        MobileAds.Initialize(this);
+        //Android.Gms.Ads.MobileAds.Initialize(this); // for ads
+
+        var googlePlayAuthenticationService = new GooglePlayAuthenticationService(this, AssemblyInfo.FirebaseClientId);
+        var authenticationService = new AppAuthenticationService(AssemblyInfo.StoreBaseUri, AssemblyInfo.StoreAppId, googlePlayAuthenticationService, AssemblyInfo.IsDebugMode);
+        var googlePlayBillingService = GooglePlayBillingService.Create(this);
+        VpnHoodApp.Instance.AccountService = new AppAccountService(authenticationService, googlePlayBillingService);
+    }
+
+    protected override void OnDestroy()
+    {
+        VpnHoodApp.Instance.AccountService = null;
+        base.OnDestroy();
     }
 
 }