+++ VpnHood.Client.App.WebServer/Api/IBillingController.cs
@@ -1,0 +1,11 @@
+﻿using System.Diagnostics.CodeAnalysis;
+using VpnHood.Client.App.Abstractions;
+
+namespace VpnHood.Client.App.WebServer.Api;
+
+[SuppressMessage("ReSharper", "UnusedMember.Global")]
+public interface IBillingController
+{
+    Task<SubscriptionPlan[]> GetSubscriptionPlans();
+    Task Purchase(string userId, string planId);
+}