+++ VpnHood.Client/Exceptions/NoStableVpnException.cs
@@ -1,8 +1,4 @@
 ﻿namespace VpnHood.Client.Exceptions;
 
-public class NoStableVpnException : Exception
-{
-    public NoStableVpnException() : base("VPN was connected, but it looked like the connection was not stable.")
-    {
-    }
-}
+public class NoStableVpnException() 
+    : Exception("VPN was connected, but it looked like the connection was not stable.");