+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v3.3.462
+### Client
+* Fix: Updating ServerTokenUrl from the server
+
+* ### Server
+* Fix: Updating ServerTokenUrl & ServerSecret from the server
+
+
 # v3.3.454
 ### Client
 * Update: New Access Token format
@@ -7,7 +15,7 @@
 * Feature: Update server token URL from the server
 * Feature: Android: Prompt for the update by Google Play if possible
 * Feature: Android: Support opening (KEY) as VpnHood Key file
-* Fix: Windows: Freeze network after auto-reconnect with muti-server
+* Fix: Windows: Freeze network after auto-reconnect redirects to a new server
 * Deprecate: Server ProtocolVersion 3 is deprecated and no longer supported
 
 ### Server